+++ tests/Moq.Tests/SequentialActionExtensionsFixture.cs
@@ -39,6 +39,36 @@ namespace Moq.Tests
 			Assert.Throws<ArgumentException>(() => mock.Object.Do());
 		}
 
+		[Fact]
+		public void PerformSequenceWithCalculatedExceptions()
+		{
+			var mock = new Mock<IFoo>();
+
+			mock.SetupSequence(m => m.Do())
+				.Pass()
+				.Throws<InvalidOperationException>(() => new InvalidOperationException())
+				.Throws(() => new ArgumentException());
+
+			mock.Object.Do();
+			Assert.Throws<InvalidOperationException>(() => mock.Object.Do());
+			Assert.Throws<ArgumentException>(() => mock.Object.Do());
+		}
+
+		[Fact]
+		public void PerformSequenceWithThrowCalculatedExceptionFirst()
+		{
+			var mock = new Mock<IFoo>();
+
+			mock.SetupSequence(m => m.Do())
+				.Throws<InvalidOperationException>(() => new InvalidOperationException())
+				.Pass()
+				.Throws(new ArgumentException());
+
+			Assert.Throws<InvalidOperationException>(() => mock.Object.Do());
+			mock.Object.Do();
+			Assert.Throws<ArgumentException>(() => mock.Object.Do());
+		}
+
 		public interface IFoo
 		{
 			void Do();